+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 #### 2.0.0-alpha1 - ??
 * Removed `ElementAt` method accepting an `IndexingStrategy` (use `Cycle` instead)
+* Renamed `CountsExactly` to `HasExactly`
 * Added `Chunk` method
 * Added `Cycle` method
 * Added `Distinct` method